+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: New protocol: Foxify (Sonic chain)
 - Add: New protocol: Liquidity Royalty Tranching (Berachain)
 - Add: New protocol: cSigma Finance (Ethereum)
+- Add: New protocol: Spark (Ethereum)
 - Fix: Various RPC error code workarounds (Monad, Arbitrum, Hyperliquid)
 
 # 0.37